Added a service layer to the OPDS system [#1307]
 * Removed auditing of OPDS requests.

Moved the logic for navigating the OPDS feeds to services, leaving
the REST layer to simply call those methods.
