+++ comixed-opds/src/main/java/org/comixedproject/opds/rest/OPDSDateController.java
@@ -18,23 +18,13 @@
 
 package org.comixedproject.opds.rest;
 
-import static org.comixedproject.opds.model.OPDSAcquisitionFeed.ACQUISITION_FEED_LINK_TYPE;
-import static org.comixedproject.opds.model.OPDSNavigationFeed.NAVIGATION_FEED_LINK_TYPE;
-import static org.comixedproject.opds.rest.OPDSLibraryController.SELF;
-import static org.comixedproject.opds.rest.OPDSLibraryController.SUBSECTION;
-
 import java.security.Principal;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.GregorianCalendar;
 import lombok.NonNull;
 import lombok.extern.log4j.Log4j2;
-import org.comixedproject.auditlog.rest.AuditableRestEndpoint;
 import org.comixedproject.opds.OPDSException;
-import org.comixedproject.opds.OPDSUtils;
 import org.comixedproject.opds.model.*;
-import org.comixedproject.service.comicbooks.ComicBookService;
+import org.comixedproject.opds.service.OPDSAcquisitionService;
+import org.comixedproject.opds.service.OPDSNavigationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -49,100 +39,43 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @Log4j2
 public class OPDSDateController {
-  private static final String STORE_DATE_YEARS_ID = "20";
-  private static final int COMIC_STORE_DATE_FOR_YEAR_ID = 20;
-
-  @Autowired private ComicBookService comicBookService;
-  @Autowired private OPDSUtils opdsUtils;
-
-  private SimpleDateFormat simpleDateFormat = new SimpleDateFormat("MMM dd yyyy");
+  @Autowired private OPDSNavigationService opdsNavigationService;
+  @Autowired private OPDSAcquisitionService opdsAcquisitionService;
 
   /**
    * Returns navigation links for the store date years in the library.
    *
+   * @param unread the unread flag
    * @return the years as navigation links
    */
   @GetMapping(value = "/opds/dates/released", produces = MediaType.APPLICATION_XML_VALUE)
-  @AuditableRestEndpoint(logResponse = true)
   @PreAuthorize("hasRole('READER')")
   @ResponseBody
-  public OPDSNavigationFeed loadYears(
+  public OPDSNavigationFeed getYearsFeed(
       @RequestParam(name = "unread", defaultValue = "false") final boolean unread) {
     log.info("Loading comic years");
-    final OPDSNavigationFeed response =
-        new OPDSNavigationFeed("Store Date: Years", STORE_DATE_YEARS_ID);
-    this.comicBookService.getYearsForComics().stream()
-        .sorted()
-        .forEach(
-            year -> {
-              log.trace("Adding year: {}", year);
-              OPDSNavigationFeedEntry entry =
-                  new OPDSNavigationFeedEntry(String.valueOf(year), String.valueOf(year));
-              entry.setContent(new OPDSNavigationFeedContent(String.format("The Year %d", year)));
-              entry
-                  .getLinks()
-                  .add(
-                      new OPDSLink(
-                          ACQUISITION_FEED_LINK_TYPE,
-                          SUBSECTION,
-                          String.format(
-                              "/opds/dates/released/years/%d/weeks?unread=%s", year, unread)));
-              response.getEntries().add(entry);
-            });
-    return response;
+    return this.opdsNavigationService.getYearsFeed(unread);
   }
 
   /**
    * Returns navigation links for the weeks of the given year in the library.
    *
    * @param year the year
+   * @param unread the unread flag
    * @return the weeks as navigation links
    * @throws OPDSException if an error occurs
    */
   @GetMapping(
       value = "/opds/dates/released/years/{year}/weeks",
       produces = MediaType.APPLICATION_XML_VALUE)
-  @AuditableRestEndpoint(logResponse = true)
   @PreAuthorize("hasRole('READER')")
   @ResponseBody
-  public OPDSNavigationFeed loadWeeksForYear(
+  public OPDSNavigationFeed getWeeksFeedForYear(
       @PathVariable("year") @NonNull final Integer year,
       @RequestParam(name = "unread", defaultValue = "false") final boolean unread)
       throws OPDSException {
     log.info("Loading comics for year {}", year);
-    final OPDSNavigationFeed response =
-        new OPDSNavigationFeed(
-            "Comics For Year: " + year, String.valueOf(COMIC_STORE_DATE_FOR_YEAR_ID + year));
-    log.trace("Loading days with comics for year");
-    this.comicBookService.getWeeksForYear(year).stream()
-        .sorted()
-        .forEach(
-            weekNumber -> {
-              log.trace("Adding week {} of {}", weekNumber, year);
-              final Date weekStarts = this.getDateFor(year, weekNumber, Calendar.SUNDAY);
-              final Date weekEnds = this.getDateFor(year, weekNumber, Calendar.SATURDAY);
-              OPDSNavigationFeedEntry entry =
-                  new OPDSNavigationFeedEntry(
-                      String.format("Year %d Week %d", year, weekNumber),
-                      String.valueOf(year + weekNumber));
-              entry.setContent(
-                  new OPDSNavigationFeedContent(
-                      String.format(
-                          "%s thru %s",
-                          this.simpleDateFormat.format(weekStarts),
-                          this.simpleDateFormat.format(weekEnds))));
-              entry
-                  .getLinks()
-                  .add(
-                      new OPDSLink(
-                          ACQUISITION_FEED_LINK_TYPE,
-                          SUBSECTION,
-                          String.format(
-                              "/opds/dates/released/years/%d/weeks/%d?unread=%s",
-                              year, weekNumber, String.valueOf(unread))));
-              response.getEntries().add(entry);
-            });
-    return response;
+    return this.opdsNavigationService.getWeeksFeedForYear(year, unread);
   }
 
   /**
@@ -150,12 +83,12 @@ public class OPDSDateController {
    *
    * @param year the year
    * @param week the week
+   * @param unread the unread flag
    * @return the comics as acquisition links
    */
   @GetMapping(
       value = "/opds/dates/released/years/{year}/weeks/{week}",
       produces = MediaType.APPLICATION_XML_VALUE)
-  @AuditableRestEndpoint(logResponse = true)
   @PreAuthorize("hasRole('READER')")
   @ResponseBody
   public OPDSAcquisitionFeed loadComicsForYearAndWeek(
@@ -163,39 +96,8 @@ public class OPDSDateController {
       @PathVariable("year") @NonNull final Integer year,
       @PathVariable("week") @NonNull final Integer week,
       @RequestParam(name = "unread", defaultValue = "false") final boolean unread) {
-    log.info("Loading comics for year {}", year);
-    final Date weekStarts = this.getDateFor(year, week, Calendar.SUNDAY);
-    final Date weekEnds = this.getDateFor(year, week, Calendar.SATURDAY);
-
-    final OPDSAcquisitionFeed response =
-        new OPDSAcquisitionFeed(
-            String.format(
-                "Comics For Week Of %s To %s",
-                simpleDateFormat.format(weekStarts), simpleDateFormat.format(weekEnds)),
-            String.valueOf(COMIC_STORE_DATE_FOR_YEAR_ID + year));
-    log.trace("Loading comics");
-    this.comicBookService.getComicsForYearAndWeek(year, week, principal.getName(), unread).stream()
-        .forEach(
-            comicBook -> {
-              log.trace("Adding comic to collection entries: {}", comicBook.getId());
-              response.getEntries().add(this.opdsUtils.createComicEntry(comicBook));
-            });
-    response
-        .getLinks()
-        .add(
-            new OPDSLink(
-                NAVIGATION_FEED_LINK_TYPE,
-                SELF,
-                String.format(
-                    "/opds/dates/released/years/%d/weeks/%d?unread=%s", year, week, unread)));
-    return response;
-  }
-
-  private Date getDateFor(final Integer year, final Integer week, final int dayOfWeek) {
-    final GregorianCalendar calendar = new GregorianCalendar();
-    calendar.set(Calendar.YEAR, year);
-    calendar.set(Calendar.WEEK_OF_YEAR, week);
-    calendar.set(Calendar.DAY_OF_WEEK, dayOfWeek);
-    return calendar.getTime();
+    final String email = principal.getName();
+    log.info("Loading comics for year {} and week {} for {}", year, week, email);
+    return this.opdsAcquisitionService.getComicsFeedForYearAndWeek(email, year, week, unread);
   }
 }
